Removed data object from FlightsList component
Now the data is coming from a file exporting the object. Now codeclimate shouldnt trow errors.
